Problem 1
So i have to make a cylindrical fuel tank for go kart. Now we only know the our required volume but do not know the combination of radius and height which will be best suited to get the desired volume. So reduce my work i had written this code in which we will enter the maximum radius, maximum height of the cylinder and volume required. As a result we will get the desired combination.