GO KART FUEL TANK PROBLEM
From last few days i was working on the fuel tank for Go kart in which i have to measure the fuel left in the tank. So, i used a cylindrical fuel tank, now i have to find out that pair of radius and height which is suitable for my fuel requirement.